Allow loading keypairs from a directory
Provide a means to load pre-existing keys from a directory.
(As opposed to generating public and private keys on the fly for every
certificate.)

This improves deterministic repository generation, but also
significantly speeds up the process, because generating the keys is the
slowest step by far.

One way to generate the keys is

	mkdir "keypairs"
	for i in $(seq 0 100); do
		echo $i
		openssl genrsa -out "keypairs/$i.pem" 2048
	done

Then fire up barry like this:

	barry --keys "keypairs/" ... potato.rd

The key names need to be numbered monotonically from zero to whatever
number of keys "potato.rd" needs, minus one.