Use regex validations in functional tests
I don't really want to upload the test keychains to the git repository;
they might grow into a lot of clutter as the test suite grows.
So I'd rather regenerate them whenever the tmpfs box is mounted.

Of course, this means the tests will not have consistent signatures and
hashes, so the validations need to account for it.

Plus, hardcoding exact hashes and signatures across many tests sounds
fishy. I don't think it'll scale well.

So replace the exact matching checks with regex matching checks.