Re-implement Names
Names were implemented as arrays of objects, when in reality they're
supposed to be arrays of arrays of objects.

I removed the short version (the ability to declare a name as a simple
string which will be converted to commonName) because it's too much
trouble for little gain.

This is a pending loose end from the JSON-ification commit.